fsm updates
Lil discord message:

And the idea behind the abstraction is that we have the current state and the event. I wrote a thing to get the next state. Then I have a map from current, next state -> fxn. So we execute the function and advance the state in the loop.